split the reader module into recent, historical and utils
