+++ chain/client/src/archive/mod.rs
@@ -1,3 +1,5 @@
-pub mod cloud_archival_reader;
+pub mod cloud_archival_utils;
 pub mod cloud_archival_writer;
+pub mod cloud_historical_reader;
+pub mod cloud_recent_reader;
 pub mod cold_store_actor;
